dwm init

+++ install.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/env bash
-set -e
+# set -e
 
 PWDIR="$( cd -- "$(dirname "$0")" >/dev/null 2>&1 ; pwd -P )"
 
@@ -10,8 +10,29 @@ BLUE="\e[3;34m"
 GRAY="\e[3;37m"
 ENDCOLOR="\e[0m"
 
+
+HOSTNAME="arch"
+PLATFORM="AMD"
+USERNAME="syaofox"
+PASSWORD="0928"
+SWAPSIZE="17408"
+
+apptitle="Arch Linux Fast Install"
+GITNAME="syaofox"
+GITMAIL="syaofox@gmail.com"
+
+USERHOME="/home/${USERNAME}"
+REPODIR="${USERHOME}/Repos"
+
+SUCKLESSDIR="${REPODIR}/suckless"
+
 tip() {
     echo -e "${GREEN}${1:-Done!}${ENDCOLOR}"
+}
+
+pressanykey(){
+	# read -n1 -p "${txtpressanykey}"
+	read -n1 -p "$(echo -e $GREEN${txtpressanykey}$ENDCOLOR)"
 }
 
 ptip() {
@@ -24,36 +45,165 @@ etip() {
     read -p "$(echo -e $YELLOW${2:-Press enter to continue...}$ENDCOLOR)" temp
 }
 
+
+
+source ${PWDIR}/scripts/strings.sh
 source ${PWDIR}/scripts/base.sh
+source ${PWDIR}/scripts/config.sh
+source ${PWDIR}/scripts/desktop.sh
 
-apptitle="Arch Linux Fast Install"
 
-loadstrings(){
 
-	locale=en_US.UTF-8
-	#font=
+installpkgmenu() {
+	if [ "${1}" = "" ]; then
+		nextitem="."
+	else
+		nextitem=${1}
+	fi
 
-	txtexit="Exit"
-	txtback="Back"
-	txtignore="Ignore"
+	options=()
+	options+=("${txtinstallvideodrivers}" "  ${PLATFORM}")
+	options+=("${txtinstallnetwork}" "")
+	options+=("${txtinstallsoundandblooth}" "")
+	options+=("${txtinstallutils}" "")
+	sel=$(whiptail --backtitle "${apptitle}" --title "${txtInstallPkg}" --menu "" --cancel-button "${txtexit}" --default-item "${nextitem}" 0 0 0 \
+		"${options[@]}" \
+		3>&1 1>&2 2>&3)
+	if [ "$?" = "0" ]; then
+		case ${sel} in
+			"${txtinstallvideodrivers}")
+				installvideodrivers
+				nextitem="${txtinstallnetwork}"
+			;;
+			"${txtinstallnetwork}")
+				installnetwork
+				nextitem="${txtinstallsoundandblooth}"
+			;;
+			"${txtinstallsoundandblooth}")
+				installsoundandblooth
+				nextitem="${txtinstallutils}"
+			;;
+			"${txtinstallutils}")
+				installutils
+				nextitem="${txtinstallutils}"
+			;;
+		esac
+		installpkgmenu "${nextitem}"
+	fi
+}
 
-	txtmainmenu="Main Menu"
-	txtsetdatetime="Set timedate"
-	txtsetsourcelist="Set Source List"
-	txtdiskpartmenu="Disk Partitions"
-	txteditparts="Edit Partitions"
-	txtselectpartsmenu="Select Partitions and Install"
-	txtselectdevice="Select %1 device :"
-	txtselecteddevices="Selected devices :"
 
-	txtreboot="Reboot"
+# TODO configmengu
+configmenu(){
+	if [ "${1}" = "" ]; then
+		nextitem="."
+	else
+		nextitem=${1}
+	fi
+
+	options=()
+	options+=("${txtsethostname}" "/etc/hostname")
+	options+=("${txtsetlocale}" "/etc/locale.conf, /etc/locale.gen")
+	options+=("${txtsettime}" "/etc/localtime")
+	options+=("${txtsetrootpassword}" "")
+	options+=("${txtadduser}" "  ${USERNAME}")
+	options+=("${txtInstallPkg}" "")
+	options+=("${txtconfigsamba}" "")	
+	options+=("${txtinstallgrub}" "")
+
+	sel=$(whiptail --backtitle "${apptitle}" --title "${txtarchinstallmenu}" --menu "" --cancel-button "${txtexit}" --default-item "${nextitem}" 0 0 0 \
+		"${options[@]}" \
+		3>&1 1>&2 2>&3)
+	if [ "$?" = "0" ]; then
+		case ${sel} in			
+			"${txtsethostname}")
+				sethostname
+				nextitem="${txtsetlocale}"
+			;;
+			"${txtsetlocale}")
+				setlocale				
+				nextitem="${txtsettime}"
+			;;
+			"${txtsettime}")
+				setlocaltime				
+				nextitem="${txtsetrootpassword}"
+			;;
+			"${txtsetrootpassword}")
+				setrootpasswordchroot				
+				nextitem="${txtadduser}"
+			;;
+			"${txtadduser}")
+				adduser				
+				nextitem="${txtInstallPkg}"
+			;;
+			"${txtInstallPkg}")
+				installpkgmenu			
+				nextitem="${txtconfigsamba}"
+			;;
+			"${txtconfigsamba}")
+				configsamba			
+				nextitem="${txtinstallgrub}"
+			;;
+			"${txtinstallgrub}")
+				installgrub			
+				nextitem="${txtinstallgrub}"
+			;;
+	esac
+		configmenu "${nextitem}"
+	fi
+}
+
+# TODO installmenu
+installmenu(){
+	if [ "${1}" = "" ]; then
+		nextitem="${txtinstallarchlinux}"
+	else
+		nextitem=${1}
+	fi
+	options=()
+	options+=("${txtselectmirrorsbycountry}" "(${txtoptional})")
+	options+=("${txteditmirrorlist}" "(${txtoptional})")
+	
+	options+=("${txtinstallarchlinux}" "pacstrap")
+	options+=("${txtgenfstab}" "/etc/fstab")
+	options+=("${txtconfigarchlinux}" "")
+	sel=$(whiptail --backtitle "${apptitle}" --title "${txtinstallmenu}" --menu "" --cancel-button "${txtunmount}" --default-item "${nextitem}" 0 0 0 \
+		"${options[@]}" \
+		3>&1 1>&2 2>&3)
+	if [ "$?" = "0" ]; then
+		case ${sel} in
+			"${txtselectmirrorsbycountry}")
+				selectmirrorsbycountry
+				nextitem="${txtinstallarchlinux}"
+			;;
+			"${txteditmirrorlist}")
+				${EDITOR} /etc/pacman.d/mirrorlist
+				nextitem="${txtinstallarchlinux}"
+			;;
+			"${txtinstallarchlinux}")
+				if(installbase) then
+					nextitem="${txtgenfstab}"
+				fi
+			;;
+			"${txtgenfstab}")
+				archgenfstab
+				nextitem="${txtconfigarchlinux}"
+			;;
+			"${txtconfigarchlinux}")
+				switchtochroot
+				nextitem="${txtconfigarchlinux}"
+			;;
+		esac
+		installmenu "${nextitem}"
+	else
+		unmountdevices
+	fi
 }
 
 
 
 
-
-
+# TODO mainmenu
 mainmenu(){
 	if [ "${1}" = "" ]; then
 		nextitem="."
@@ -62,7 +212,6 @@ mainmenu(){
 	fi
 	options=()
 	options+=("${txtsetdatetime}" "")	
-	options+=("${txtsetsourcelist}" "")	
 	options+=("${txtdiskpartmenu}" "")	
 	options+=("${txtselectpartsmenu}" "")	
 	options+=("" "")
@@ -74,11 +223,7 @@ mainmenu(){
 		case ${sel} in
 			"${txtsetdatetime}")
 				settimedate
-				nextitem="${txtreboot}"
-			;;
-			"${txtsetsourcelist}")
-				setsourcelist
-				nextitem="${txtreboot}"
+				nextitem="${txtdiskpartmenu}"
 			;;
 			"${txtdiskpartmenu}")
 				diskpartcgdisk
@@ -99,16 +244,54 @@ mainmenu(){
 	fi
 }
 
-if [ "${chroot}" = "1" ]; then
-	case ${command} in
-		'setrootpassword') archsetrootpasswordchroot;;
-    esac
+
+
+
+# TODO main
+
+
+
+if [ `whoami` = root ]; then
+
+	while (( "$#" )); do
+		case ${1} in	
+			--chroot) 
+				chroot=1
+				command=${2}
+				args=${3}
+				
+			;;
+		esac
+		shift
+	done
+
+
+	if [ "${chroot}" = "1" ]; then
+		case ${command} in
+			'config') 
+
+				systemctl stop reflector
+				pacman -S --noconfirm --needed arch-install-scripts wget libnewt vim	
+				loadstrings
+				EDITOR=vim
+				configmenu
+				exit
+			;;
+		esac
+	else
+		systemctl stop reflector
+		pacman -S --noconfirm --needed arch-install-scripts wget libnewt vim	
+		loadstrings
+		EDITOR=vim
+		mainmenu
+	fi
+
 else
-	systemctl stop reflector
-	pacman -S --needed arch-install-scripts wget libnewt vim	
+	sudo pacman -Syy
+	sudo pacman -S --noconfirm --needed arch-install-scripts wget libnewt vim	
 	loadstrings
 	EDITOR=vim
-	mainmenu
+	desktopmenu
 fi
 
 exit 0
